requirements.txt
tensorflow>=2.15.0
numpy
Pillow
pycocotools
matplotlib
tqdm
